Support config rollback across container restarts

+++ docker-entrypoint.sh
@@ -18,7 +18,11 @@ _init() {
     if [ ! -L "$_origin" ] && [ -d "$_origin" ] && [ ! -e "$_defaults" ]; then
         cp -R "$_origin" "$_defaults"
     fi
-    rm -rf "$_link"
+
+    if [ ! -L "$_link" ]; then
+        # Only delete "/etc/nginx" if it's not a symlink (but a directory containing defaults)
+        rm -rf "$_link"
+    fi
 }
 
 _prepare_nginx_cache_dir /var/cache/nginx www-data:www-data
